Added user roles. Added names for routes.

+++ app/Http/Controllers/CompanyController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use \App\Models\Company;
+use \App\Models\Company\Campaign;
 use \App\Models\User;
 use \App\Models\UserCompany;
 use \App\Rules\CompanyWebhookActionsRule;
@@ -11,6 +13,13 @@ use Validator;
 
 class CompanyController extends Controller
 {
+    /**
+     * List all companies
+     * 
+     * @param Request $request
+     * 
+     * @return Response
+     */
     public function list(Request $request)
     {
         $limit  = intval($request->limit) ?: 25;
@@ -38,11 +47,22 @@ class CompanyController extends Controller
         ]);
     }
 
+    /**
+     * Create a company
+     * 
+     * @param Request $request
+     * 
+     * @return Response
+     */
     public function create(Request $request)
     {
         $rules = [
             'name'            => 'required|max:255',
-            'webhook_actions' => ['required', 'json', new CompanyWebhookActionsRule()]
+            'webhook_actions' => [
+                'required', 
+                'json', 
+                new CompanyWebhookActionsRule()
+            ]
         ];
 
         $validator = Validator::make($request->input(), $rules);
@@ -56,6 +76,7 @@ class CompanyController extends Controller
 
         $company = Company::create([
             'account_id'        => $user->account_id,
+            'created_by'        => $user->id,
             'name'              => $request->name,
             'webhook_actions'   => $request->webhook_actions
         ]);
@@ -66,6 +87,14 @@ class CompanyController extends Controller
         ], 201);
     }
 
+    /**
+     * View a company
+     * 
+     * @param Request $request
+     * @param Company $company
+     * 
+     * @return Response
+     */
     public function read(Request $request, Company $company)
     {
         return response([
@@ -74,11 +103,23 @@ class CompanyController extends Controller
         ]);
     }
 
+    /**
+     * Update a company
+     * 
+     * @param Request $request
+     * @param Company $company
+     * 
+     * @return Response
+     */
     public function update(Request $request, Company $company)
     {
         $rules = [
             'name'            => 'required|max:255',
-            'webhook_actions' => ['required', 'json', new CompanyWebhookActionsRule()]
+            'webhook_actions' => [
+                'required', 
+                'json', 
+                new CompanyWebhookActionsRule()
+            ]
         ];
         
         $validator = Validator::make($request->input(), $rules);
@@ -98,19 +139,29 @@ class CompanyController extends Controller
         ]);
     }
 
+    /**
+     * Delete a company
+     * 
+     * @param Request $request
+     * @param Company $company 
+     * 
+     * @return Response
+     */
     public function delete(Request $request, Company $company)
     {
-        //
-        //  TODO: Make sure no records are attached
-        //  ...
-        //
+        if( $company->isInUse() ){
+            return response([
+                'error' => 'Company has active campaigns attached'
+            ], 400);
+        }        
 
-        //  Remove from users
-        User::where('company_id', $company->id)
-             ->update(['company_id' => null]);
+        //  Remove any inactive campaigns attached
+        Campaign::where('company_id', $company->id)
+                ->delete();
 
+        //  Remove any users attached
         UserCompany::where('company_id', $company->id)
-                    ->delete();
+                   ->delete();
 
         $company->delete();
